requirements.txt
Jinja2==2.10.1
SQLAlchemy==1.3.7
mailcleaner-config==0.3
PyMySQL==0.9.3
pytest==4.4.0
pytest-cov==2.7.1
coverage==4.5.3
factory-boy==2.12.0
yapf==0.18.0